format currentPath and move utils to format-path
